- clarify finance service

+++ api/src/main/java/ds2/oss/core/api/FinanceServices.java
@@ -9,10 +9,10 @@ public interface FinanceServices {
     /**
      * Calculates a netto value based on the given brutto value.
      *
-     * @param val         the value
+     * @param val         the brutto value
      * @param vat         usually 119 for a vat value of 19 percent
      * @param targetScale the target scale
      * @return the calculated price
      */
-    BigDecimal calculateNettoPrice(BigDecimal val, long vat, int targetScale);
+    BigDecimal calculateNettoPrice(BigDecimal val, BigDecimal vat, int targetScale);
 }
